Finish basic moving with turn order enforced
So now:
- you can select any unit YOU own
- you can command it to move to a space within its legal move range
- you can NOT select any unit you DON'T own
- you can NOT move your selected unit to a space it can't move to (this
includes spaces other units are in, as well)
- you can deselect by clicking anywhere that isn't one of your units
- turn order is automatically handled by a script that checks if any
legal plays are left to be made. if there are none, it passes turn. we
had discussed a button before, I am totally okay with that design
decision and am not trying to comment to the contrary with this commit,
it was just easier to get done and I didn't want to worry about a button
yet. haha.
- units will refresh on turn end. their color and movability will be
restored, though if they are not controlled by the person whose turn it
is they will not be selectable.
- I left a hook in for rounds ending. Doesn't do much of anything now,
but I was thinking forward about mana and stuff.